+++ finagle-zipkin-scribe/src/test/scala/com/twitter/finagle/zipkin/thrift/ZipkinTracerTest.scala
@@ -1,7 +1,7 @@
 package com.twitter.finagle.zipkin.thrift
 
 import com.twitter.app.Flags
-import com.twitter.finagle.zipkin.core.{DefaultSampler, Sampler}
+import com.twitter.finagle.zipkin.core.Sampler
 import com.twitter.finagle.zipkin.initialSampleRate
 import org.scalatest.FunSuite
 
@@ -17,7 +17,6 @@ class ZipkinTracerTest extends FunSuite {
   test("passing sample rate in the ZipkinTracer constructor always takes precedence") {
     val flag = new Flags("foo", true)
     flag.parseArgs(Array("-com.twitter.finagle.zipkin.initialSampleRate", "0.63"))
-    DefaultSampler.reset()
 
     val tracer = new ZipkinTracer(ScribeRawZipkinTracer(), 1.0f)
     assert(tracer.getSampleRate == 1.0f)